Move code for filling in slides to utils, and update tests accordingly.
